opennetem/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name="opennetem"
version = "0.2"
dependencies = [
	"pandas",
	"certifi==2024.2.2",
	"charset-normalizer==3.3.2",
	"docker==7.0.0",
	"et-xmlfile==1.1.0",
	"idna==3.6",
	"influxdb-client==1.40.0",
	"numpy==1.26.4",
	"openpyxl==3.1.2",
	"packaging==23.2",
	"pandas==2.2.0",
	"python-dateutil==2.8.2",
	"pytz==2024.1",
	"reactivex==4.0.4",
	"requests==2.31.0",
	"six==1.16.0",
	"statsd-telegraf==3.2.1.post1",
	"typing_extensions==4.9.0",
	"tzdata==2024.1",
	"urllib3==2.2.1"
]

[project.scripts]
opennetem = "opennetem.scenario:opennetem"
on_mon_rtt = "opennetem.on_mon_rtt:do_main"
on_bpstats = "opennetem.on_bpstats:do_main"
on_bplist =  "opennetem.on_bplist:do_main"

